requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
decorator==5.1.1
frozenlist==1.3.3
fsspec==2022.11.0
idna==3.4
importlib-metadata==5.2.0
joblib==1.2.0
librosa==0.9.2
lightning-utilities==0.4.2
llvmlite==0.39.1
multidict==6.0.3
numba==0.56.4
numpy==1.23.5
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pkg_resources==0.0.0
pooch==1.6.0
protobuf==3.20.1
pycparser==2.21
python-dateutil==2.8.2
pytorch-lightning==1.8.6
pytz==2022.7
PyYAML==6.0
requests==2.28.1
resampy==0.4.2
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
soundfile==0.11.0
tensorboardX==2.5.1
threadpoolctl==3.1.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchmetrics==0.11.0
torchvision==0.14.1+cu116
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
zipp==3.11.0
